pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pymupdf>=1.25.5",
    "pytest>=8.3.5",
]
